refactor(console): combine steps in connector guide
