+++ packages/console/src/components/Markdown/index.module.scss
@@ -1,21 +1,60 @@
+@use '@/scss/underscore' as _;
+
 .markdown {
-  font: var(--font-body-medium);
+  li {
+    font: var(--font-body-medium);
+
+    ul,
+    ol {
+      padding-inline-start: 1ch;
+    }
+  }
+
+  ul {
+    padding-inline-start: 4ch;
+
+    > li {
+      margin-block-start: _.unit(2);
+      margin-block-end: _.unit(2);
+      padding-inline-start: _.unit(1);
+    }
+  }
+
+  ol {
+    padding-inline-start: 2ch;
+
+    > li {
+      margin-block-start: _.unit(3);
+      margin-block-end: _.unit(3);
+      padding-inline-start: _.unit(1);
+    }
+  }
+
+  a {
+    font: var(--font-body-medium);
+    color: var(--color-text-link);
+  }
 
   h1 {
     font: var(--font-title-large);
+    margin: _.unit(6) 0;
   }
 
   h2 {
     font: var(--font-title-medium);
-  }
-
-  h3,
-  h4,
-  h5 {
-    font: var(--font-title-small);
+    color: var(--color-caption);
+    margin: _.unit(6) 0 _.unit(3);
   }
 
   p {
     font: var(--font-body-medium);
+    margin: _.unit(3) 0;
   }
+}
+
+.inlineCode {
+  background: var(--color-layer-2);
+  font: var(--font-body-medium);
+  padding: _.unit(1) _.unit(1);
+  border-radius: 4px;
 }
